pyproject.toml
[tool.poetry]
name = "pavai"
version = "0.1.0"
description = "private ai assistant"
authors = ["mychen76 <mychen76@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "pavai/vocei_web",from = "bases"},
    {include = "pavai/talkie_cli",from = "bases"},
    {include = "pavai/finetune_api",from = "bases"},
    {include = "pavai/solar",from = "components"},
    {include = "pavai/finetune",from = "components"},
    {include = "pavai/chatcommand",from = "components"},
    {include = "pavai/shared",from = "components"},
    {include = "pavai/helloworld",from = "components"},
    {include = "pavai/translator",from = "components"},    
]

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.7.0"
gradio = "^4.15.0"
torch = {version = "^2.1.2+cu118", source = "pytorch"}
torchvision = {version = "^0.16.2+cu118", source = "pytorch"}
torchaudio = {version = "^2.1.2+cu118", source = "pytorch"}
python-dotenv = "^1.0.0"
halo = "^0.0.31"
gitpython = "^3.1.41"
psutil = "^5.9.8"
retrying = "^1.3.4"
sounddevice = "^0.4.6"
pyaudio = "^0.2.14"
webrtcvad = "^2.0.10"
duckduckgo-search = "^4.2"
pytube = "^15.0.0"
tiktoken = "^0.5.2"
onnxruntime = "^1.16.3"
huggingface-hub = "^0.20.3"
faster-whisper = "^0.10.0"
gtts = "^2.5.0"
phonemizer = "^3.2.1"
espeak-phonemizer = "^1.3.1"
openai = "^1.9.0"
transformer-utils = "^0.1.1"
diffusers = {extras = ["torch"], version = "^0.25.1"}
sentence-transformers = "^2.2.2"
transformers = "^4.37.0"
llama-cpp-python = "^0.2.32"
soundfile = "^0.12.1"
munch = "^4.0.0"
pydub = "^0.25.1"
pyyaml = "^6.0.1"
librosa = "^0.10.1"
nltk = "^3.8.1"
accelerate = "^0.26.1"
einops = "^0.7.0"
einops-exts = "^0.0.4"
tqdm = "^4.66.1"
typing-extensions = "^4.9.0"
monotonic-align = {git = "https://github.com/resemble-ai/monotonic_align.git"}
instructor = "^0.4.7"
spacy = "^3.7.2"
presidio-analyzer = "^2.2.352"
presidio-anonymizer = "^2.2.352"
html2text = "^2020.1.16"
markdown = "^3.5.2"
pymupdf = "^1.23.17"
docx2txt = "^0.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
jupyterlab = "^4.0.11"
black = "^23.12.1"
faker = "^22.4.0"
ipykernel = "^6.29.0"
pylint = "^3.0.3"
isort = "^5.13.2"
jupyterlab-code-formatter = "^2.2.1"

[tool.poetry.group.test.dependencies]
faker = "^22.4.0"

[tool.poetry.group.finetune.dependencies]
fastapi = "^0.109.0"
pymongo = "^4.6.1"
uvicorn = "^0.26.0"
pika = "^1.3.2"

[tool.poetry.group.extras.dependencies]
redis = "^5.0.1"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
